+++ mex/editor/client/src/app/app.ts
@@ -1,13 +1,12 @@
-import { Component, ChangeDetectionStrategy } from "@angular/core";
-import { MatIcon } from "@angular/material/icon";
+import { Component } from "@angular/core";
 import { RouterOutlet } from "@angular/router";
 import { NavBarComponent } from "./components/nav-bar-component/nav-bar-component";
+import { TranslocoDirective } from "@jsverse/transloco";
 
 @Component({
   selector: "app-root",
-  imports: [RouterOutlet, MatIcon, NavBarComponent],
+  imports: [RouterOutlet, NavBarComponent, TranslocoDirective],
   templateUrl: "./app.html",
-  changeDetection: ChangeDetectionStrategy.Eager,
   styleUrl: "./app.scss",
 })
 export class App {}