+++ mex/editor/client/src/test-providers.ts
@@ -1,7 +1,34 @@
 import { provideHttpClient, withXhr } from "@angular/common/http";
 import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { provideRouter } from "@angular/router";
+import { TranslocoTestingModule } from "@jsverse/transloco";
+import { translocoConfig, translocoProviders } from "./app/transloco";
 
-const testProviders = [provideRouter([]), provideHttpClient(withXhr()), provideHttpClientTesting()];
+
+/* eslint-disable @typescript-eslint/naming-convention */
+const de = {
+  "test.headline": "Titelzeile",
+  "test.icuMessageFormat": "{count, plural, one {EINER} other {VIELE}}",
+};
+
+const en = {
+  "test.headline": "Headline",
+  "test.icuMessageFormat": "{count, plural, one {ONE} other {MANY}}",
+};
+/* eslint-enable @typescript-eslint/naming-convention */
+
+const [_, ...rest] = translocoProviders;
+
+const testProviders = [
+  provideRouter([]),
+  provideHttpClient(withXhr()),
+  provideHttpClientTesting(),
+  TranslocoTestingModule.forRoot({
+    langs: { de, en },
+    translocoConfig,
+    preloadLangs: true,
+  }).providers,
+  rest,
+];
 
 export default testProviders;