+++ mex/editor/main.py
@@ -12,6 +12,7 @@ from starlette.exceptions import HTTPException as StarletteHTTPException
 
 from mex.common.logging import logger
 from mex.editor.api.backend import router as backend_router
+from mex.editor.api.i18n import router as i18n_router
 from mex.editor.api.system import router as system_router
 from mex.editor.frontend import STATIC_DIR, npm_watch
 from mex.editor.logging import UVICORN_LOGGING_CONFIG
@@ -44,6 +45,9 @@ class SPAStaticFiles(StaticFiles):
             return await super().get_response("index.html", scope)
 
 
+API_PREFIX = "/api/v0"
+
+
 @asynccontextmanager
 async def dev_lifespan(_: FastAPI) -> AsyncGenerator[None]:  # pragma: no cover
     """Run npm watch during dev mode."""
@@ -71,8 +75,9 @@ def create_fastapi(
         allow_headers=["*"],
     )
     if startup in ["api", "both"]:
-        app.include_router(backend_router, prefix="/api/v0")
-        app.include_router(system_router, prefix="/api/v0")
+        app.include_router(backend_router, prefix=API_PREFIX)
+        app.include_router(system_router, prefix=API_PREFIX)
+        app.include_router(i18n_router, prefix=API_PREFIX)
     if startup in ["frontend", "both"]:
         app.mount(
             "/",